+++ z0bug_odoo/z0bug_odoo/egg-info/CHANGELOG.rst
@@ -1,3 +1,10 @@
+2.0.13.1 (2023-12-01)
+~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] TestEnv: test on model asset.asset
+* [FIX] TestEnv: minor fixes in mixed environment excel + zerobug
+* [FIX] TestEnv: sometimes external.KEY did not work
+
 2.0.13 (2023-12-01)
 ~~~~~~~~~~~~~~~~~~~